+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 
 * Added workaround for an issue where a postgresql cache of a coco dataset is unable to remember its original json filepath.
+* Fixed issue with `CocoSqlDatabase.disconnect` where a session would be dereferenced, but not closed.
 
 
 ## Version 0.7.8 - Released 2024-03-20